Boss Charge Attack Added
ฝากแมพลองดูหน่อย ตัวบอสมันไม่ขยับอะ หลังจากใช้สกิล

+++ 2D Shooting Project/Assets/Scripts/Enemy/DamageInput.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DamageInput : MonoBehaviour
 {
-    public int damage;
+    public float damage;
 
     private void OnCollisionEnter2D(Collision2D collider)
     {
